+++ source/core/src/core/contracts/process_manager/PersistSubmittedTransaction/persist_submitted_transaction_proto_version.py
@@ -1,0 +1,2 @@
+class PersistSubmittedTransactionProtoVersion:
+    version_1 = "1"