+++ source/core/src/core/silver/entry_points.py
@@ -1,4 +1,5 @@
 import core.silver.application.streams.calculated_stream as calculated_stream
+import core.silver.application.streams.migrated_transactions as migrated_transactions
 import core.silver.application.streams.notify_transactions_persisted_stream as notify_transactions_persisted_stream
 import core.silver.application.streams.submitted_transactions as submitted_transactions
 
@@ -13,3 +14,7 @@ def stream_submitted_transactions() -> None:
 
 def notify_transactions_persisted() -> None:
     notify_transactions_persisted_stream.notify()
+
+
+def stream_migrated_transactions() -> None:
+    migrated_transactions.stream_migrated_transactions_to_silver()