+++ docs/Measurements.Client/ReleaseNotes/ReleaseNotes.md
@@ -1,12 +1,24 @@
 # Measurements.Client Release Notes
 
+## Version 1.4.0
+
+- Updated measurement points to only contain latest values.
+
+## Version 1.3.0
+
+- Added method to get measurements based on period.
+
+## Version 1.2.0
+
+- Updated Client interface.
+
 ## Version 1.1.4
 
 - Bugfix. Don't reference Measurements from Client.
 
 ## Version 1.1.3
 
-- Bugfix. Deserializing MeasurementDto
+- Bugfix. Deserializing MeasurementDto.
 
 ## Version 1.1.2